Add Variant A and B on Portfolio Carousel
